minutely precipitation intensity overview

+++ app/src/main/java/wangdaye/com/geometricweather/weather/converters/CaiyunResultConverter.java
@@ -480,7 +480,7 @@ public class CaiyunResultConverter {
                                     currentWeatherCode
                             ),
                             1,
-                            null,
+                            (Integer) null,
                             null
                     )
             );
